feat(querier): implement the Tempo gRPC querier protocol
Replace the SignalDBQuerier stub (which returned fabricated data or
panicked via unimplemented!) with a real tempopb.Querier implementation
backed by TraceService, served on the querier's Flight port so a Tempo
query-frontend can use SignalDB as a querier:

- FindTraceByID converts the internal trace model to OTLP-shaped
  ResourceSpans grouped per service
- SearchRecent maps Tempo search requests onto SignalDB search params
  and returns TraceSearchMetadata with spans_per_span_set honored
- Tenant resolution follows an authenticated TenantContext when the
  internal auth interceptor is configured, else Tempo's X-Scope-OrgID
  header, else default/default
- SearchBlock returns Unimplemented (SignalDB has no Tempo block model);
  tag endpoints advertise the statically queryable tag set instead of
  fabricated values

Querier-internal SearchQueryParams time/duration fields widen to i64 so
Tempo's u32 epoch seconds and ms durations map losslessly (ticket JSON
stays wire-compatible). Monolithic mode is unchanged: the protocol
targets standalone querier deployments in Tempo clusters.

+++ src/querier/src/query/mod.rs
@@ -22,9 +22,13 @@ pub struct FindTraceByIdParams {
 pub struct SearchQueryParams {
     pub q: Option<String>,
     pub tags: Option<String>,
-    pub min_duration: Option<i32>,
-    pub max_duration: Option<i32>,
+    /// Minimum span duration in nanoseconds
+    pub min_duration: Option<i64>,
+    /// Maximum span duration in nanoseconds
+    pub max_duration: Option<i64>,
     pub limit: Option<i32>,
-    pub start: Option<i32>,
-    pub end: Option<i32>,
+    /// Search window start (unix seconds)
+    pub start: Option<i64>,
+    /// Search window end (unix seconds)
+    pub end: Option<i64>,
 }
